Former-commit-id: b6156081301561be5ed00f3b2cddc324744b53ce

+++ Sample/source/TimeStepTracker.h
@@ -14,23 +14,23 @@ class TimeStepGuard;
 //and are smoothed over 5 recommendations
 class TimeStepTracker {
 	//The shortest goal timestep
-	float timeStepMinGoal;
+	double timeStepMinGoal;
 	//the longest acceptable timestep
-	float timeStepMaxGoal;
+	double timeStepMaxGoal;
 	//used to smooth recommendations
-	MovingAverage<float> recommendationAverage;
+	MovingAverage<double> recommendationAverage;
 
 	//Add a recommendation to the average based off of 
 	//the amount of time it took to do whatever it is this tracker is tracking
 	//added by a TimeStepGuard
-	void addRecommendation(float timeStepTook);
+	void addRecommendation(double timeStepTook);
 
 	friend class TimeStepGuard;
 public:
-	TimeStepTracker(float minGoal, float maxGoal);
+	TimeStepTracker(double minGoal, double maxGoal);
 
 	//Get the recommendation from .5 to 2
 	// <1 indicates a speed up is required (decrease detail of render/update)
 	// >1 indicates a slow down is ok (increase detail of render/update)
-	float GetRecommendation();
+	double GetRecommendation();
 };
